Update reinstall_cuda.sh
Added installation check to verify success.

+++ ubuntu-build/reinstall_cuda.sh
@@ -50,5 +50,10 @@ printf "\n************************* INSTALLING NVIDIA-GDS **********************
 sudo apt install -y nvidia-gds
 printf "\n****************** INSTALLATION COMPLETE - CLEANING UP *******************\n\n"
 sudo apt autoremove -y
-printf "\n******************************* REBOOTING ********************************\n\n"
-sudo reboot
+printf "\n**************************** CHECKING INSTALL ****************************\n\n"
+if [[ $(nvidia-smi|sed -n 3p|grep "Driver Version"|wc -l) -eq 0 ]] then
+  printf "\nINSTALL DID NOT COMPLETE. CHECK FOR ERRORS.\n\n"
+else
+  printf "\n******************************* REBOOTING ********************************\n\n"
+  sudo reboot
+fi
